requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
scikit-learn>=1.4.0
scipy>=1.11.0
openai>=1.30.0
yfinance>=0.2.40
pypdf>=4.2.0
pdfplumber>=0.10.0
pymupdf>=1.24.0
openpyxl>=3.1.0
python-docx>=1.1.0

